+++ internal/octogo/parser.go
@@ -6352,12 +6352,14 @@ state1:
 //		| float_lit
 //		| string_lit
 //		| rune_lit
-//		| "(" Expression ")"
-//		| "[" [ Expression ] "]" Type [ CompositeLit ]
+//		| "(" Expression ")" [ FactorSuffix ]
+//		| "[" [ Expression ] "]" Type [ CompositeLit [ FactorSuffix ] ]
 //		| "chan" Type
 //		| FuncLiteral [ FactorSuffix ] .
 //
 //	State 0
+//		on  identifier
+//			shift and goto state 10
 //		on  "chan"
 //			shift and goto state 1
 //		on  float_lit, int_lit, rune_lit, string_lit
@@ -6365,11 +6367,9 @@ state1:
 //		on  '('
 //			shift and goto state 3
 //		on  '['
-//			shift and goto state 5
-//		on  identifier
-//			shift and goto state 9
+//			shift and goto state 6
 //		on  "func"
-//			call FuncLiteral and goto state 10
+//			call FuncLiteral and goto state 5
 //	State 1
 //		on  "chan", "func", "interface", "struct", '*', '[', identifier
 //			call Type and goto state 2
@@ -6380,32 +6380,36 @@ state1:
 //			call Expression and goto state 4
 //	State 4
 //		on  ')'
-//			shift and goto state 2
+//			shift and goto state 5
 //	State 5
-//		on  ']'
-//			shift and goto state 6
-//		on  "<-", "chan", "func", '!', '&', '(', '*', '+', '-', '[', '^', '~', float_lit, identifier, int_lit, rune_lit, string_lit
-//			call Expression and goto state 8
-//	State 6
-//		on  "chan", "func", "interface", "struct", '*', '[', identifier
-//			call Type and goto state 7
-//	State 7
-//		Accept
-//		on  '{'
-//			call CompositeLit and goto state 2
-//	State 8
-//		on  ']'
-//			shift and goto state 6
-//	State 9
-//		Accept
-//		on  '{'
-//			call CompositeLit and goto state 2
-//		on  '(', '.', '['
-//			call FactorSuffix and goto state 7
-//	State 10
 //		Accept
 //		on  '(', '.', '['
 //			call FactorSuffix and goto state 2
+//	State 6
+//		on  ']'
+//			shift and goto state 7
+//		on  "<-", "chan", "func", '!', '&', '(', '*', '+', '-', '[', '^', '~', float_lit, identifier, int_lit, rune_lit, string_lit
+//			call Expression and goto state 9
+//	State 7
+//		on  "chan", "func", "interface", "struct", '*', '[', identifier
+//			call Type and goto state 8
+//	State 8
+//		Accept
+//		on  '{'
+//			call CompositeLit and goto state 5
+//	State 9
+//		on  ']'
+//			shift and goto state 7
+//	State 10
+//		Accept
+//		on  '{'
+//			call CompositeLit and goto state 2
+//		on  '(', '.', '['
+//			call FactorSuffix and goto state 11
+//	State 11
+//		Accept
+//		on  '{'
+//			call CompositeLit and goto state 2
 //
 // Factor is used internally from Parse.
 func (p *Parser) Factor() (r []int32) {
@@ -6414,6 +6418,9 @@ func (p *Parser) Factor() (r []int32) {
 	// state0:
 	accept, errorSet = false, 61
 	switch Symbol(p.tok.Ch) {
+	case identifier:
+		r = append(r, p.shift())
+		goto state10
 	case TOK_chan:
 		r = append(r, p.shift())
 		goto state1
@@ -6425,13 +6432,10 @@ func (p *Parser) Factor() (r []int32) {
 		goto state3
 	case TOK_005b:
 		r = append(r, p.shift())
-		goto state5
-	case identifier:
-		r = append(r, p.shift())
-		goto state9
+		goto state6
 	case TOK_func:
 		r = p.add(r, p.FuncLiteral())
-		goto state10
+		goto state5
 	}
 	return p.stop(r, accept, errorSet)
 state1:
@@ -6458,45 +6462,53 @@ state4:
 	switch Symbol(p.tok.Ch) {
 	case TOK_0029:
 		r = append(r, p.shift())
-		goto state2
+		goto state5
 	}
 	return p.stop(r, accept, errorSet)
 state5:
+	accept, errorSet = true, 82
+	switch Symbol(p.tok.Ch) {
+	case TOK_0028, TOK_002e, TOK_005b:
+		r = p.add(r, p.FactorSuffix())
+		goto state2
+	}
+	return p.stop(r, accept, errorSet)
+state6:
 	accept, errorSet = false, 27
 	switch Symbol(p.tok.Ch) {
 	case TOK_005d:
 		r = append(r, p.shift())
-		goto state6
+		goto state7
 	case TOK_003c002d, TOK_chan, TOK_func, TOK_0021, TOK_0026, TOK_0028, TOK_002a, TOK_002b, TOK_002d, TOK_005b, TOK_005e, TOK_007e, float_lit, identifier, int_lit, rune_lit, string_lit:
 		r = p.add(r, p.Expression())
-		goto state8
+		goto state9
 	}
 	return p.stop(r, accept, errorSet)
-state6:
+state7:
 	accept, errorSet = false, 59
 	switch Symbol(p.tok.Ch) {
 	case TOK_chan, TOK_func, TOK_interface, TOK_struct, TOK_002a, TOK_005b, identifier:
 		r = p.add(r, p.Type())
-		goto state7
+		goto state8
 	}
 	return p.stop(r, accept, errorSet)
-state7:
+state8:
 	accept, errorSet = true, 108
 	switch Symbol(p.tok.Ch) {
 	case TOK_007b:
 		r = p.add(r, p.CompositeLit())
-		goto state2
+		goto state5
 	}
 	return p.stop(r, accept, errorSet)
-state8:
+state9:
 	accept, errorSet = false, 107
 	switch Symbol(p.tok.Ch) {
 	case TOK_005d:
 		r = append(r, p.shift())
-		goto state6
+		goto state7
 	}
 	return p.stop(r, accept, errorSet)
-state9:
+state10:
 	accept, errorSet = true, 83
 	switch Symbol(p.tok.Ch) {
 	case TOK_007b:
@@ -6504,14 +6516,14 @@ state9:
 		goto state2
 	case TOK_0028, TOK_002e, TOK_005b:
 		r = p.add(r, p.FactorSuffix())
-		goto state7
+		goto state11
 	}
 	return p.stop(r, accept, errorSet)
-state10:
-	accept, errorSet = true, 82
+state11:
+	accept, errorSet = true, 108
 	switch Symbol(p.tok.Ch) {
-	case TOK_0028, TOK_002e, TOK_005b:
-		r = p.add(r, p.FactorSuffix())
+	case TOK_007b:
+		r = p.add(r, p.CompositeLit())
 		goto state2
 	}
 	return p.stop(r, accept, errorSet)